requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.1
django-paypal==1.1.2
gunicorn==20.1.0
idna==3.3
Pillow==9.0.0
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
django-paypal==1.1.2